Rename LocalLocations file
